requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alabaster==0.7.16
albumentations==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
arabic-reshaper==3.0.0
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
bleach==6.1.0
boto3==1.35.49
botocore==1.35.49
build==1.2.2.post1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.0
coverage==5.3.1
cryptography==43.0.3
cssselect2==0.7.0
cycler==0.12.1
data-gradients==0.3.2
datasets==2.21.0
Deprecated==1.2.14
dill==0.3.8
docutils==0.17.1
einops==0.3.2
exceptiongroup==1.2.2
fastapi==0.115.4
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.6.1
future==1.0.0
glob2==0.7
grpcio==1.67.0
h11==0.14.0
html5lib==1.1
httptools==0.6.4
huggingface-hub==0.26.2
humanfriendly==10.0
hydra-core==1.3.2
idna==3.7
imagededup==0.3.2
imageio==2.36.0
imagesize==1.4.1
imutils==0.5.4
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json-tricks==3.16.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaggle==1.6.17
kiwisolver==1.4.7
lazy_loader==0.4
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.0
onnxruntime-gpu==1.19.2
onnxsim==0.4.36
onnxslim==0.1.35
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
oscrypto==1.3.0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pip-tools==7.4.1
platformdirs==4.3.6
propcache==0.2.0
protobuf==3.20.3
psutil==6.1.0
py-cpuinfo==9.0.0
pyarrow==18.0.0
pycocotools==2.0.8
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyDeprecate==0.3.2
Pygments==2.18.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
pyparsing==3.2.0
pypdf==5.1.0
pyproject_hooks==1.2.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
python-slugify==8.0.4
pytube==15.0.0
pytz==2024.2
PyWavelets==1.7.0
PyYAML==6.0.2
qrcode==8.0
qudida==0.0.4
RapidFuzz==3.10.1
referencing==0.35.1
reportlab==3.6.13
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.3
roboflow==1.1.48
rpds-py==0.20.0
s3transfer==0.10.3
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==4.0.3
sphinx-rtd-theme==1.3.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
starlette==0.41.2
stringcase==1.2.0
super-gradients==3.7.1
svglib==1.5.1
svgwrite==1.4.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==3.5.0
tifffile==2024.9.20
tinycss2==1.4.0
tomli==2.0.2
torch==2.5.0
torchmetrics==0.8.0
torchvision==0.20.0
tqdm==4.66.5
Tree==0.2.4
treelib==1.6.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ultralytics==8.3.23
ultralytics-thop==2.0.9
uritools==4.0.3
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
webencodings==0.5.1
websockets==13.1
Werkzeug==3.0.6
wrapt==1.16.0
xhtml2pdf==0.2.11
xxhash==3.5.0
yarl==1.17.0
